requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.4
django-exo-mentions==1.1.3
djangorestframework==3.12.2
filelock==3.0.12
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.20.1
virtualenv==20.2.2
